masukan nilai density oke

+++ app/monita.h
@@ -233,7 +233,7 @@ volatile struct t_st_hw st_hw;
 struct t_kalib {
 	float m;
 	float C;
-	short density;
+	int density;
 	char status;		// [status lihat define di atas !!!] //
 	char adc;
 };
